--cleaned up some function calls in ImGui containers, and the ShaderMachineBase class, to prevent warnings from being generated (being nice to compiler)

+++ OrganicGLWinLib/ImGuiButtonPanel.h
@@ -20,7 +20,7 @@ class ImGuiButtonPanel
 			xSize = 200;
 			ySize = 100;
 		};
-		ImGuiButtonPanel(std::string in_panelName, int in_xOffset, int in_yOffset, int in_xSize, int in_ySize) :
+		ImGuiButtonPanel(std::string in_panelName, float in_xOffset, float in_yOffset, float in_xSize, float in_ySize) :
 			panelName(in_panelName),
 			xOffset(in_xOffset),
 			yOffset(in_yOffset),
@@ -31,10 +31,10 @@ class ImGuiButtonPanel
 
 		std::unordered_map<std::string, ImGuiButton> buttonMap;
 		std::string panelName = "";
-		int xOffset = 0;
-		int yOffset = 0;
-		int xSize = 0;
-		int ySize = 0;
+		float xOffset = 0;
+		float yOffset = 0;
+		float xSize = 0;
+		float ySize = 0;
 
 		ImGuiButtonClickResult getClickResult()
 		{
