+++ OrganicGLWinLib/ImGuiSliderFloatPanel.h
@@ -19,7 +19,7 @@ class ImGuiSliderFloatPanel
 			ySize = 100;
 		};
 
-		ImGuiSliderFloatPanel(std::string in_panelName, int in_xOffset, int in_yOffset, int in_xSize, int in_ySize) :
+		ImGuiSliderFloatPanel(std::string in_panelName, float in_xOffset, float in_yOffset, float in_xSize, float in_ySize) :
 			panelName(in_panelName),
 			xOffset(in_xOffset),
 			yOffset(in_yOffset),
@@ -29,10 +29,10 @@ class ImGuiSliderFloatPanel
 
 		std::unordered_map<std::string, ImGuiSliderFloat> sliderMap;
 		std::string panelName = "";
-		int xOffset = 0;
-		int yOffset = 0;
-		int xSize = 0;
-		int ySize = 0;
+		float xOffset = 0;
+		float yOffset = 0;
+		float xSize = 0;
+		float ySize = 0;
 
 		void renderSliderFloats()
 		{
